Sun, Jun 11, 2023, 11:01 PM +00:00

+++ desmos_graph_shifter.py
@@ -13,7 +13,7 @@ def main():
         case 2:
             sign, sign_dummy, a, a_dummy, b = request_input()
             sign = sign_list(sign)
-            unchanged, adder, b_dummy = int(input('How many do you want unchanged: ')), int(input('How much do you want to shift (+x/-x): ')), ''
+            unchanged, adder, b_dummy = int(input('How many do you want unchanged: ')), float(input('How much do you want to shift (+x/-x): ')), ''
             print('\ny_{1}=',end='')
 
             for i in range(unchanged):
